feat(api) add /register and /authenticate endpoints. refactor(restructure existing code)

+++ src/main/java/com/ang/rest/domain/entities/TransactionDetailsId.java
@@ -8,6 +8,6 @@ import java.io.Serializable;
 @NoArgsConstructor
 @AllArgsConstructor
 public class TransactionDetailsId implements Serializable {
-    private Long transaction;
-    private Long product;
+    private Integer transaction;
+    private Integer product;
 }
